ref(native): Remove the legacy stacktrace processor
