+++ tests/sentry/lang/native/test_utils.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from sentry.lang.native.utils import get_sdk_from_event, cpu_name_from_data
+from sentry.lang.native.utils import get_sdk_from_event, is_minidump_event
 
 
 def test_get_sdk_from_event():
@@ -39,35 +39,63 @@ def test_get_sdk_from_event():
     assert sdk_info['version_patchlevel'] == 1
 
 
-def test_cpu_name_from_data():
-    cpu_name = cpu_name_from_data(
-        {
-            'contexts': {
-                'device': {
-                    'type': 'device',
-                    'arch': 'arm64'
-                },
-                'device2': {
-                    'type': 'device',
-                    'arch': 'arm7'
-                },
-            }
-        }
-    )
-
-    assert cpu_name == 'arm64'
-
-
-def test_cpu_name_from_data_inferred_type():
-    cpu_name = cpu_name_from_data(
-        {
-            'contexts': {
-                'some_device': {
-                    'type': 'device',
-                    'arch': 'arm64'
+def test_is_minidump():
+    assert is_minidump_event({
+        'exception': {
+            'values': [{
+                'mechanism': {
+                    'type': 'minidump'
                 }
-            }
+            }]
         }
-    )
+    })
 
-    assert cpu_name == 'arm64'
+    assert not is_minidump_event({
+        'exception': {
+            'values': [{
+                'mechanism': {
+                    'type': 'other'
+                }
+            }]
+        }
+    })
+
+    assert not is_minidump_event({
+        'exception': {
+            'values': [{
+                'mechanism': {
+                    'type': None
+                }
+            }]
+        }
+    })
+
+    assert not is_minidump_event({
+        'exception': {
+            'values': [{
+                'mechanism': None
+            }]
+        }
+    })
+
+    assert not is_minidump_event({
+        'exception': {
+            'values': [None]
+        }
+    })
+
+    assert not is_minidump_event({
+        'exception': {
+            'values': []
+        }
+    })
+
+    assert not is_minidump_event({
+        'exception': {
+            'values': None
+        }
+    })
+
+    assert not is_minidump_event({
+        'exception': None
+    })
